Update drawing-functions.js
reworked how the hover-highlighting is done, so that I'm modifying properties of the already-draw graph elements, rather than re-drawing the whole thing.
I imagine that this is cheaper/faster, but importantly it also gets rid of the problem that the mouseenter events kept being continually triggered, due to the elements being redrawn under the cursor.
Also added a function to identify and 'sink' unconnected nodes, giving them coordinates at the bottom of the graph.